rename: 타입 선언 파일의 폴더 위치 수정
